+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.1.0] - 2026-05-02
+
+### Added
+ - Added a `rule34Py.get_favorites` method.
+ - Added a `rule34Py.get_favorites_id` method.
+ - Added a logo and favicon to the docs.
+
 ## [4.0.2] - 2025-12-22
 
 ### Fixed